perf(params): Test performance and memory enhancements around params.
This is the client-side of an experiment to see what sort (if any)
performance enhancements we can get by leaving params on everything
that has params as raw byte arrays until it is absolutely required
that they be something else.

Inspired by watching the memory usage of the load of a large customer,
and noticing a giamt spike in memory usage when we got to remarshalling
everything from raw JSON into the final in-memory format.