feat(consensus): Check cluster fail over safeness.
Add a CLI command that allows us to test to see if a cluster failing
over will lose any data.  It takes an optional timeout to let any
in-flight changes get propogated to the rest of the cluster members.

If this command returns true, then there is at least one ono-observer
member that can become the active node without losing any in-flight
commits.